+++ bin/scripts/testStackTrace.php
@@ -32,7 +32,7 @@ class TestClass
     public static function myStaticMethod()
     {
         $clone = new \Awesomite\ErrorDumper\Cloners\ClonedException(new TmpException('My test exception'));
-        $clone->getStackTrace()->setVarDumper(new \Awesomite\StackTrace\VarDumpers\LightVarDumper());
+        $clone->getStackTrace()->setVarDumper(new \Awesomite\VarDumper\LightVarDumper());
 
         if (php_sapi_name() === 'cli') {
             $view = new \Awesomite\ErrorDumper\Views\ViewCli(7, 3);